@@ -5,7 +5,7 @@ author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 homepage: https://github.com/memspace/zefyr
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: '>=2.8.1 <3.0.0'
 
 dependencies:
   flutter:
@@ -13,7 +13,9 @@ dependencies:
   collection: ^1.14.6
   url_launcher: ^5.0.0
   quill_delta: ^1.0.2
-  notus: ^0.1.0
+  flutter_tex: ^3.5.2
+  notus:
+    path: ../notus
   meta: ^1.1.0
   quiver_hashcode: ^2.0.0
 
