@@ -13,24 +13,19 @@ void main() {
   runApp(ZefyrApp());
 }
 
-// Create a Focus Intent that does nothing
-class FakeFocusIntent extends Intent {
-  const FakeFocusIntent();
-}
-
 class ZefyrApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       // disable default arrows focus changes.
       // otherwise it makes the keyboard flicker when we move with arrows)
-      shortcuts: Map<LogicalKeySet, Intent>.from(WidgetsApp.defaultShortcuts)
-        ..addAll(<LogicalKeySet, Intent>{
-          LogicalKeySet(LogicalKeyboardKey.arrowLeft): const FakeFocusIntent(),
-          LogicalKeySet(LogicalKeyboardKey.arrowRight): const FakeFocusIntent(),
-          LogicalKeySet(LogicalKeyboardKey.arrowDown): const FakeFocusIntent(),
-          LogicalKeySet(LogicalKeyboardKey.arrowUp): const FakeFocusIntent(),
-        }),
+//      shortcuts: Map<LogicalKeySet, Intent>.from(WidgetsApp.defaultShortcuts)
+//        ..addAll(<LogicalKeySet, Intent>{
+//          LogicalKeySet(LogicalKeyboardKey.arrowLeft): const FakeFocusIntent(),
+//          LogicalKeySet(LogicalKeyboardKey.arrowRight): const FakeFocusIntent(),
+//          LogicalKeySet(LogicalKeyboardKey.arrowDown): const FakeFocusIntent(),
+//          LogicalKeySet(LogicalKeyboardKey.arrowUp): const FakeFocusIntent(),
+//        }),
       debugShowCheckedModeBanner: false,
       title: 'Zefyr Editor',
       home: HomePage(),
