@@ -161,8 +161,10 @@ class InputConnectionController implements TextInputClient {
     }
   }
 
-  @override
-  AutofillScope get currentAutofillScope => null; //AutofillGroup.of(context);
+  // re-add once flutter 1.18 is released
+  //@override
+  //AutofillScope get currentAutofillScope => null; //AutofillGroup.of(context);
+
   @override
   void showAutocorrectionPromptRect(int start, int end) {
     // TODO: implement showAutocorrectionPromptRect
